Added map make and C++20 features

+++ TGTools/TGToolsConsole/TGTools.cpp
@@ -109,12 +109,12 @@ int main(int count, const char** args) {
 			std::vector<std::string> results(std::istream_iterator<std::string>{iss},
 				std::istream_iterator<std::string>());
 
-			const int count = results.size();
+			const size_t count = results.size();
 			args = new const char* [count];
 			for (size_t i = 0; i < count; i++)
 				args[i] = results[i].c_str();
 
-			const Result result = exec(count, args);
+			const Result result = exec((int)count, args);
 			if (result != Result::SUCCESS)
 				std::cout << "Error (" << (int)result << ")!" << std::endl;
